Extract auto refresh control to isolate typing re-renders
